Refactor code, and move functions to subfield6Utils.js
